Fixing format of HVM memory block and agent so Long-Mem-Eval can properly function
