+++ asdrp/memory/__init__.py
@@ -1,1 +1,1 @@
-from .hierarchical_vector_memory import HierarchicalVectorMemory
+from .hvm import HierarchicalVectorMemory
